Authentication: implement a basic form of Authentication
This commit introduces a new endpoint, '/authenticate', which can only
be accessed via a POST request, with JSON arguments as
{
    'username': username,
    'password': password
}.
If the request succeeds, then a JSON response will be returned, with a key
arguments that represents the authentication token.
ALL AUTHENTICATION TOKENS ARE - with the current state - TEMPORARY
This means that whenever the server is restarted, all tokens are lost and
users need to login again.

From now on, every API call needs to have a GET argument: 'auth_key' that represents
the current token. If this argument is not included, the request will stop immediately.